Add mutation validation code

+++ src/graphql/mutation.ts
@@ -5,11 +5,12 @@ import { z } from "zod";
 //creates a zod schema that validates a title field
 const todoSchema = z.object({
   title: z.string().min(1, "Title is required").max(255, "Title is too long"),
+  id: z.string().length(10, "ID should look like: 62c98cea-4474-4161-9933-8d4e45db3069"),
+  completed: z.boolean(),
 });
 
 export const Mutation: IMutation<Context> = {
   createSomething: async (_, { input }, { prisma }) => {
-    //validation
    
     const something = await prisma.something.create({
       data: {
@@ -24,12 +25,13 @@ export const Mutation: IMutation<Context> = {
   },
 
   createTodo: async (_, { input }, { prisma }) => {
+    //validation
     const validation = todoSchema.safeParse(input);
     if (!validation.success){
       throw new Error(validation.error.errors.map(err => err.message).join(", "))
     }
 
-
+    //creating to do
     const todo = await prisma.todo.create({
       data: {
         title: input.title,
@@ -45,6 +47,13 @@ export const Mutation: IMutation<Context> = {
   },
 
   changeTodoStatus: async (_, { id, completed }, { prisma }) => {
+    //validation
+    const validation = todoSchema.safeParse({id, completed});
+    if (!validation.success){
+      throw new Error(validation.error.errors.map(err => err.message).join(", "))
+    }
+
+    //updating
     const updatedTodo = await prisma.todo.update({
       where: {
         id: id,
@@ -62,6 +71,13 @@ export const Mutation: IMutation<Context> = {
   },
 
   changeTodoTitle: async (_, { id, title }, { prisma }) => {
+    //validation
+    const validation = todoSchema.safeParse({id, title});
+    if (!validation.success){
+      throw new Error(validation.error.errors.map(err => err.message).join(", "))
+    }
+
+    //updating
     const updatedTodo = await prisma.todo.update({
       where: {
         id: id,
@@ -79,6 +95,13 @@ export const Mutation: IMutation<Context> = {
   },
 
   deleteTodo: async (_, { id }, { prisma }) => {
+    //validation
+    const validation = todoSchema.safeParse(id);
+    if (!validation.success){
+      throw new Error(validation.error.errors.map(err => err.message).join(", "))
+    }
+
+    //deleting
     await prisma.todo.delete({
       where: {
         id: id,
